+++ src/funcs/petMyPet123.ts
@@ -81,6 +81,7 @@ export async function petMyPet123(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "POST",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,